fix: add codec interface
