+++ protocol/dubbo3/impl/package.go
@@ -1,0 +1,14 @@
+package impl
+
+type TripleRequestHeader struct {
+}
+
+type TripleHeader struct {
+}
+
+type TriplePackage struct {
+	header *TripleHeader
+	codec  *CodeC
+}
+
+NewTriplePkgHandler
